Fixed alignment of 'desc' in add_note() function.
The ELF spec specifically lists the alignment of the namez char[] to be
4 bytes. To quote it:

"Padding is present, if necessary, to ensure 4-byte alignment for the
descriptor. Such padding is not included in namesz."

However, the current implementation sets the alignment to either 4 or 8
bytes depending on the class of the ELF file (CLASS32 or CLASS64). This
commit fixes the alignment to only 4 bytes in all cases.